move web reader content loading into viewmodel
