+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/root/RootView.kt
@@ -84,6 +84,8 @@ fun PrimaryNavigator(
     }
 
     composable("WebReader/{slug}") {
+      webReaderViewModel.reset() // clear previously loaded item
+
       WebReaderLoadingContainer(
         it.arguments?.getString("slug") ?: "",
         webReaderViewModel = webReaderViewModel
